Format columns with decimal alignment with constraints only once
Instead of formatting the cells in order to compute their widths, then
constraining the maximum width, then formatting the cells again for
display, we now only format the cells once, in full, and pass along the
width constraints.

By doing so, we have more knowledge about the column width at format
time, and we're able to align decimals.

Closes #87