Adds new extend-rows flag (-e) to disable truncation (#116)
Adds a new extend-rows flag (-e) which allows the user to disable row
truncation. This is super helpful for wide csv files that are larger
than the terminal width. When using this flag, pipe output to `less -S`
to be able to use the arrow keys to scroll right and left to actually
view the overflowing columns.

Eg:

    `tidy-viewer -e wide.csv | less -S`

Or even... to preserve colors through less...

    `tidy-viewer -a -e wide.csv | less -S`.